process application args before creating the context
This is necessary because the profile corresponding to the --env
argument must be created before creating the application context so that
the application-{profile}.properties files can be used.  The consequence
is that the application arguments will have to be constructed twice,
once before the context is created and twice in the spring application
run method.  This is not a lot of work duplicated so it should not
matter too much.